+++ src/modules/_dataAccess/daSendGold.js
@@ -1,0 +1,7 @@
+import sendgold from '../app/trade/sendgold';
+import $dataAccess from './$dataAccess';
+import fallback from './fallbacks/sendgold';
+
+export default function daSendGold(playerName, amount) {
+  return $dataAccess(sendgold, fallback, playerName, amount);
+}
