+++ src/modules/app/trade/sendgold.js
@@ -1,0 +1,10 @@
+import trade from './trade';
+
+export default function sendgold(playerName, amount) {
+  return trade({
+    subcmd: 'sendgold',
+    xc: window.ajaxXC,
+    target_username: playerName,
+    gold_amount: amount,
+  });
+}
