+++ src/modules/world/newMap/doSendGold.js
@@ -1,31 +1,24 @@
-import indexAjaxDoc from '../../ajax/indexAjaxDoc';
+import daSendGold from '../../_dataAccess/daSendGold';
 import sendEvent from '../../analytics/sendEvent';
-import infoBox from '../../common/infoBox';
 import { defFetchPlayerStats } from '../../support/constants';
 import getValue from '../../system/getValue';
 import setValue from '../../system/setValue';
 import { getSendGoldOnWorld } from './sendGoldPref';
 
-function doneSendGold(doc) {
-  const info = infoBox(doc);
-  if (info === 'You successfully sent gold!' || info === '') {
-    setValue(
-      'currentGoldSentTotal',
-      parseInt(getValue('currentGoldSentTotal'), 10) + parseInt(getValue('goldAmount'), 10),
-    );
-    GameData.fetch(defFetchPlayerStats);
-  }
+function doneSendGold() {
+  setValue(
+    'currentGoldSentTotal',
+    parseInt(getValue('currentGoldSentTotal'), 10) + parseInt(getValue('goldAmount'), 10),
+  );
+  GameData.fetch(defFetchPlayerStats);
 }
 
 export default async function doSendGold() { // jQuery
   if (!getSendGoldOnWorld()) return;
   sendEvent('NewMap', 'doSendGold');
-  const doc = await indexAjaxDoc({
-    cmd: 'trade',
-    subcmd: 'sendgold',
-    xc: window.ajaxXC,
-    target_username: $('#HelperSendTo').html(),
-    gold_amount: $('#HelperSendAmt').html().replace(/[^\d]/g, ''),
-  });
-  if (doc) doneSendGold(doc);
+  const response = await daSendGold(
+    $('#HelperSendTo').html(),
+    $('#HelperSendAmt').html().replace(/[^\d]/g, ''),
+  );
+  if (response.s) doneSendGold();
 }
